new: modularized navbar
