+++ src/partials/navbar/navbarInfo.js
@@ -1,0 +1,24 @@
+const navItems = [
+  {
+    id: 1,
+    navItem: "about",
+    navItemUrl: "/#about",
+  },
+  {
+    id: 2,
+    navItem: "project",
+    navItemUrl: "/#project",
+  },
+  {
+    id: 3,
+    navItem: "contact",
+    navItemUrl: "/#contact",
+  },
+  {
+    id: 4,
+    navItem: "resume",
+    navItemUrl: "https://farhan2077.github.io/",
+  },
+];
+
+export { navItems };
